+++ src/test/demo05/Colliders.cs
@@ -21,7 +21,7 @@ namespace test
                     lge = new LittleGameEngine(winSize, "Colliders", Color.White);
                     lge.ShowColliders(Color.Red);
                     lge.onMainUpdate = OnMainUpdate;
-                    lge.busyWait = true;
+                    lge.waitMode = LittleGameEngine.WAITMODE_BUSY;
 
                     // cargamos los recursos que usaremos
                     lge.LoadImage("fondo", resourceDir + "/images/Backgrounds/FreeTileset/Fondo.png", false, false);